Add files via upload
This project deals with the comparison of Machine Learning Algorithms used for Sentiment Analysis. The dataset comprises of reviews on female clothes. First, data was loaded, missing values were handled and the data was cleaned. After that data analysis and visualization were performed to aid in sentiment analysis. The dataset contains mostly positive reviews, the number of negative ones is comparatively lower, while neutral reviews are the smallest group. 

The first goal was to use binary division into positive and negative reviews. First, I used Naive Bayes but it the result was not impressive as  it did not predict any negative reviews. As a result, I decided to utilize a Recurrent Neural Network with LSTM and GRU memory units. Both NNs showed a reliable and accurate predictions with accuracy around 92%. However, the recall of class 0 in both of them is around 66%, which means they work worse to predict negative reviews than positive reviews. We can possibly enhance the results with the custom learning rate or other additions.

The second goal was to predict the sentiment for a multi class classification that divided the dataset into three groups: positive, negative and neutral. As a result, the neural networks have the accuracy around 78%, but the recall for negative and especially neutral reviews is low. It can be caused by the difficulty of identifying 'neutral' words and the lengths of reviews can be too short to correctly identify this type of reviews.

Finally, the trained models were tested on the real data taken from the women clothes website. All in all, the best accuracy was achieved with positive reviews in all three models, while negative and especially neutral was showing the considerably worth result due to the number of provided data in the original dataset and biased wordings especially for neural reviews. 

Below you can see the notebook with the source code and all comments as well as the notebook file that you can download.